+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * Improve numeric casts [#580](https://github.com/icsharpcode/CodeConverter/issues/580)
 * Add ref to conversion of RaiseEvent where needed [#584](https://github.com/icsharpcode/CodeConverter/issues/584)
 * Rename clashing type memvers [#420](https://github.com/icsharpcode/CodeConverter/issues/420)
+* Fix conversion for string implicitly converted to enum [#476](https://github.com/icsharpcode/CodeConverter/issues/476)
 
 ### C# -> VB